♻️ refactor: Detail 리팩토링
