+++ src/components/detail/Comment/Comment.tsx
@@ -1,6 +1,6 @@
 import * as S from "./style";
-import ProfileImage from "../../../common/ProfileImage";
-import theme from "../../../../styles/theme";
+import ProfileImage from "../../common/ProfileImage";
+import theme from "../../../styles/theme";
 import { useNavigate } from "react-router-dom";
 interface CommentInterface {
   key: string;
