+++ src/components/detail/Comment/index.ts
