+++ src/components/detail/PostBody/index.ts
