+++ src/components/detail/PostBody/style.ts
