+++ src/components/detail/PostFooter/index.ts
