+++ src/components/detail/PostFooter/style.ts
