+++ src/components/detail/PostHeader/PostHeader.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import PostTitle from "./PostTitle";
+import PostTitle from "../PostTitle";
 import ProfileImage from "../../common/ProfileImage/index";
-import PostSummary from "./PostSummary";
+import PostSummary from "../PostSummary";
 import LikeBtn from "../../common/LikeBtn";
 import { ILike, IUser } from "src/types/model";
 import { ReactComponent as BackIcon } from "../../../assets/icons/icon_back.svg";
