+++ src/components/detail/PostHeader/index.ts
