+++ src/components/detail/PostHeader/style.ts
