+++ src/components/detail/PostSummary/index.ts
