+++ src/components/detail/PostSummary/style.ts
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import theme from "../../../../styles/theme";
+import theme from "../../../styles/theme";
 
 export const PostSummary = styled.table`
   width: 100%;
