+++ src/components/detail/PostTitle/PostTitle.tsx
