+++ src/components/detail/PostTitle/index.ts
