+++ src/components/detail/PostTitle/style.ts
