+++ src/components/detail/TextareaBox/index.ts
