+++ src/components/detail/TextareaBox/style.ts
