+++ src/components/detail/index.ts
@@ -1,0 +1,17 @@
+import PostHeader from "./PostHeader";
+import PostTitle from "./PostTitle";
+import PostSummary from "./PostSummary";
+import PostBody from "./PostBody";
+import PostFooter from "./PostFooter";
+import TextareaBox from "./TextareaBox";
+import Comment from "./Comment";
+
+export {
+  PostHeader,
+  PostTitle,
+  PostSummary,
+  PostBody,
+  PostFooter,
+  TextareaBox,
+  Comment
+};
