--- src/components/detail/index.tsx
@@ -1,1 +1,0 @@
-export {};
