+++ src/pages/detail/index.ts
